+++ src/main/java/de/sakurajin/evenbetterarcheology/api/datagen/generationType.java
@@ -1,0 +1,11 @@
+package de.sakurajin.evenbetterarcheology.api.datagen;
+
+public enum generationType {
+    CUBE,
+    PILLAR,
+    SUSPICIOUS,
+    STAIRS,
+    SLAB,
+    FROM_PARENT,
+
+}